made some minor refactoring
