+++ SeleniumAutomationGenerator/ComponentFactorySteps/AppenderCreatorStep.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Core;
 using Core.Models;
 using Core.Utils;
