+++ SeleniumAutomationGenerator/ComponentFactorySteps/CustomElementAttributeAppenderStep.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using Core;
+﻿using Core;
 using Core.Models;
 using System.Linq;
 using MoreLinq;
@@ -8,7 +7,7 @@ namespace SeleniumAutomationGenerator.ComponentFactorySteps
 {
     public class CustomElementAttributeAppenderStep : IComponentFactoryAppenderStep
     {
-        IElementAttributeContainer _container;
+        readonly IElementAttributeContainer _container;
         public CustomElementAttributeAppenderStep(IElementAttributeContainer container)
         {
             _container = container;
@@ -26,13 +25,6 @@ namespace SeleniumAutomationGenerator.ComponentFactorySteps
         {
             return rootElement.AutoAttributes
                 .Any(_container.ContainsCustomAttribute) && parent != null;
-        }
-
-        private void RunAppender(IComponentFileCreator parentClassCreator, IComponentClassAppender appender,
-            AutoElementData element)
-        {
-            if (parentClassCreator != null)
-                appender.AppendToClass(parentClassCreator, element);
         }
     }
 }
