+++ SeleniumAutomationGenerator/ProjectGenerator.cs
@@ -6,7 +6,6 @@ using SeleniumAutomationGenerator.Builders;
 using System;
 using Core.Utils;
 using SeleniumAutomationGenerator.BaseGeneratedClasses;
-using Core;
 
 namespace SeleniumAutomationGenerator
 {
@@ -15,10 +14,8 @@ namespace SeleniumAutomationGenerator
         public void GenerateProject(string webAppBaseDirectory, string projectName, string solutionName = null,
             string distDirectory = null)
         {
-            
             distDirectory = distDirectory ?? Environment.CurrentDirectory + "\\test";
             CreateSolution(projectName);
-            ComponentsContainer container = ComponentsContainer.Instance;
             ComponentsFactory componentsFactory = new ComponentsFactory();
             BuiltInComponentsInserter.InsertBuiltInComponents(componentsFactory);
             WebFolderToCsFilesConverter converter = new WebFolderToCsFilesConverter(componentsFactory, new HtmlFinder());
